requirements.txt
streamlit==1.33.0
pandas>=1.3.0,<2.0.0
requests>=2.28.0,<3.0.0
matplotlib>=3.5.0,<4.0.0
Pillow>=9.0.0,<11.0.0
python-dotenv>=1.0.0,<2.0.0
